Editor: Extract snackbars into a separate component
Editor snackbars are no rendered in the InterfaceSkeleton footer to avoid z-index context issues.
